@@ -6,10 +6,13 @@
   "directories": {
     "lib": "./lib"
   },
+  "files": [
+    "lib/"
+  ],
   "scripts": {
-    "eslint": "eslint .",
-    "test": "mocha test/index.js",
-    "test-cov": "istanbul cover --print both _mocha -- test/index.js",
+    "eslint": "eslint lib test",
+    "test": "mocha test/scripts/**/*.js",
+    "test-cov": "nyc npm run test",
     "jsdoc": "jsdoc --configure .jsdoc.json"
   },
   "repository": "hexojs/warehouse",
@@ -30,17 +33,17 @@
     "lodash": "^4.17.10"
   },
   "devDependencies": {
-    "chai": "^3.5.0",
+    "chai": "^4.2.0",
     "chai-as-promised": "^7.1.1",
-    "eslint": "^5.3.0",
+    "eslint": "^6.1.0",
     "eslint-config-hexo": "^3.0.0",
-    "istanbul": "^0.4.2",
     "jsdoc": "^3.4.0",
     "minami": "^1.1.1",
     "mocha": "^6.0.2",
+    "nyc": "^14.1.1",
     "sinon": "^7.1.1"
   },
   "engines": {
-    "node": "6 || 8 || >=10"
+    "node": ">=8.6.0"
   }
 }
