@@ -1,27 +1,26 @@
 'use strict';
 
 const SchemaType = require('../schematype');
-const util = require('../util');
 const ValidationError = require('../error/validation');
 
 /**
  * Boolean schema type.
- *
- * @class
- * @param {String} name
- * @param {Object} [options]
- *   @param {Boolean} [options.required=false]
- *   @param {Boolean|Function} [options.default]
- *   @param {String} [options.encoding=hex]
- * @extends {SchemaType}
  */
-function SchemaTypeBuffer(name, options) {
-  SchemaType.call(this, name, Object.assign({
-    encoding: 'hex'
-  }, options));
-}
+class SchemaTypeBuffer extends SchemaType {
 
-util.inherits(SchemaTypeBuffer, SchemaType);
+  /**
+   * @param {string} name
+   * @param {object} [options]
+   *   @param {boolean} [options.required=false]
+   *   @param {boolean|Function} [options.default]
+   *   @param {string} [options.encoding=hex]
+   */
+  constructor(name, options) {
+    super(name, Object.assign({
+      encoding: 'hex'
+    }, options));
+  }
+}
 
 /**
  * Casts data.
@@ -34,8 +33,8 @@ SchemaTypeBuffer.prototype.cast = function(value_, data) {
   const value = SchemaType.prototype.cast.call(this, value_, data);
 
   if (value == null || Buffer.isBuffer(value)) return value;
-  if (typeof value === 'string') return new Buffer(value, this.options.encoding);
-  if (Array.isArray(value)) return new Buffer(value);
+  if (typeof value === 'string') return Buffer.from(value, this.options.encoding);
+  if (Array.isArray(value)) return Buffer.from(value);
 };
 
 /**
@@ -78,7 +77,7 @@ SchemaTypeBuffer.prototype.compare = (a, b) => {
  * @return {Boolean}
  */
 SchemaTypeBuffer.prototype.parse = function(value, data) {
-  return value ? new Buffer(value, this.options.encoding) : value;
+  return value ? Buffer.from(value, this.options.encoding) : value;
 };
 
 /**
