@@ -8,7 +8,7 @@ describe('SchemaTypeBuffer', () => {
   const type = new SchemaTypeBuffer('test');
 
   it('cast()', () => {
-    const buf = new Buffer([97, 98, 99]);
+    const buf = Buffer.from([97, 98, 99]);
 
     type.cast(buf).should.eql(buf);
     type.cast(buf.toString('hex')).should.eql(buf);
@@ -16,14 +16,14 @@ describe('SchemaTypeBuffer', () => {
   });
 
   it('cast() - custom encoding', () => {
-    const buf = new Buffer([97, 98, 99]);
+    const buf = Buffer.from([97, 98, 99]);
     const type = new SchemaTypeBuffer('test', {encoding: 'base64'});
 
     type.cast(buf.toString('base64')).should.eql(buf);
   });
 
   it('cast() - default', () => {
-    const buf = new Buffer([97, 98, 99]);
+    const buf = Buffer.from([97, 98, 99]);
     const type = new SchemaTypeBuffer('test', {default: buf});
 
     type.cast().should.eql(buf);
@@ -34,7 +34,7 @@ describe('SchemaTypeBuffer', () => {
   }
 
   it('validate()', () => {
-    type.validate(new Buffer([97, 98, 99])).should.eql(new Buffer([97, 98, 99]));
+    type.validate(Buffer.from([97, 98, 99])).should.eql(Buffer.from([97, 98, 99]));
     shouldThrowError(1);
     shouldThrowError('foo');
     shouldThrowError([]);
@@ -49,42 +49,42 @@ describe('SchemaTypeBuffer', () => {
   });
 
   it('match()', () => {
-    type.match(new Buffer([97, 98, 99]), new Buffer([97, 98, 99])).should.be.true;
-    type.match(new Buffer([97, 98, 99]), new Buffer([97, 98, 100])).should.be.false;
-    type.match(undefined, new Buffer([97, 98, 99])).should.be.false;
+    type.match(Buffer.from([97, 98, 99]), Buffer.from([97, 98, 99])).should.be.true;
+    type.match(Buffer.from([97, 98, 99]), Buffer.from([97, 98, 100])).should.be.false;
+    type.match(undefined, Buffer.from([97, 98, 99])).should.be.false;
     type.match(undefined, undefined).should.be.true;
   });
 
   it('compare()', () => {
-    type.compare(new Buffer([97, 98, 99]), new Buffer([97, 98, 99])).should.eql(0);
-    type.compare(new Buffer([97, 98, 99]), new Buffer([97, 98, 100])).should.lt(0);
-    type.compare(new Buffer([97, 98, 99]), new Buffer([97, 98, 98])).should.gt(0);
-    type.compare(new Buffer([97, 98, 99])).should.eql(1);
-    type.compare(undefined, new Buffer([97, 98, 99])).should.eql(-1);
+    type.compare(Buffer.from([97, 98, 99]), Buffer.from([97, 98, 99])).should.eql(0);
+    type.compare(Buffer.from([97, 98, 99]), Buffer.from([97, 98, 100])).should.lt(0);
+    type.compare(Buffer.from([97, 98, 99]), Buffer.from([97, 98, 98])).should.gt(0);
+    type.compare(Buffer.from([97, 98, 99])).should.eql(1);
+    type.compare(undefined, Buffer.from([97, 98, 99])).should.eql(-1);
     type.compare().should.eql(0);
   });
 
   it('parse()', () => {
-    const buf = new Buffer([97, 98, 99]);
+    const buf = Buffer.from([97, 98, 99]);
     type.parse(buf.toString('hex')).should.eql(buf);
     should.not.exist(type.parse());
   });
 
   it('parse() - custom encoding', () => {
     const type = new SchemaTypeBuffer('name', {encoding: 'base64'});
-    const buf = new Buffer([97, 98, 99]);
+    const buf = Buffer.from([97, 98, 99]);
     type.parse(buf.toString('base64')).should.eql(buf);
   });
 
   it('value()', () => {
-    const buf = new Buffer([97, 98, 99]);
+    const buf = Buffer.from([97, 98, 99]);
     type.value(buf).should.eql(buf.toString('hex'));
     should.not.exist(type.value());
   });
 
   it('value() - custom encoding', () => {
     const type = new SchemaTypeBuffer('name', {encoding: 'base64'});
-    const buf = new Buffer([97, 98, 99]);
+    const buf = Buffer.from([97, 98, 99]);
     type.value(buf).should.eql(buf.toString('base64'));
   });
 });
