@@ -6,37 +6,37 @@ const fs = require('graceful-fs');
 const Model = require('./model');
 const Schema = require('./schema');
 const SchemaType = require('./schematype');
-const util = require('./util');
 const WarehouseError = require('./error');
 const pkg = require('../package.json');
 const JsonReadable = require('json-stream-stringify');
 
-/**
- * Database constructor.
- *
- * @class
- * @param {object} [options]
- *   @param {number} [options.version=0] Database version
- *   @param {string} [options.path] Database path
- *   @param {function} [options.onUpgrade] Triggered when the database is upgraded
- *   @param {function} [options.onDowngrade] Triggered when the database is downgraded
- */
-function Database(options) {
-  this.options = Object.assign({
-    version: 0,
-    onUpgrade() {},
+class Database {
 
-    onDowngrade() {}
-  }, options);
+  /**
+   * Database constructor.
+   *
+   * @param {object} [options]
+   *   @param {number} [options.version=0] Database version
+   *   @param {string} [options.path] Database path
+   *   @param {function} [options.onUpgrade] Triggered when the database is upgraded
+   *   @param {function} [options.onDowngrade] Triggered when the database is downgraded
+   */
+  constructor(options) {
+    this.options = Object.assign({
+      version: 0,
+      onUpgrade() {},
 
-  this._models = {};
+      onDowngrade() {}
+    }, options);
 
-  const _Model = this.Model = function(name, schema) {
-    Model.call(this, name, schema);
-  };
+    this._models = {};
 
-  util.inherits(_Model, Model);
-  _Model.prototype._database = this;
+    class _Model extends Model {}
+
+    this.Model = _Model;
+
+    _Model.prototype._database = this;
+  }
 }
 
 /**
