@@ -2,19 +2,21 @@
 
 const cloneDeep = require('lodash/cloneDeep');
 
-/**
- * Document constructor.
- *
- * @class
- * @param {object} data
- */
-function Document(data) {
-  if (data) {
-    const keys = Object.keys(data);
+class Document {
 
-    for (let i = 0, len = keys.length; i < len; i++) {
-      const key = keys[i];
-      this[key] = data[key];
+  /**
+   * Document constructor.
+   *
+   * @param {object} data
+   */
+  constructor(data) {
+    if (data) {
+      const keys = Object.keys(data);
+
+      for (let i = 0, len = keys.length; i < len; i++) {
+        const key = keys[i];
+        this[key] = data[key];
+      }
     }
   }
 }
