@@ -1,30 +1,26 @@
 'use strict';
 
-const util = require('./util');
+class WarehouseError extends Error {
 
-/**
- * WarehouseError constructor
- *
- * @class
- * @param {String} msg
- * @param {String} code
- * @extends Error
- */
-function WarehouseError(msg, code) {
-  Error.call(this);
+  /**
+   * WarehouseError constructor
+   *
+   * @param {string} msg
+   * @param {string} code
+   */
+  constructor(msg, code) {
+    super(msg);
 
-  if (Error.captureStackTrace) {
-    Error.captureStackTrace(this);
-  } else {
-    this.stack = new Error().stack;
+    if (Error.captureStackTrace) {
+      Error.captureStackTrace(this);
+    } else {
+      this.stack = new Error().stack;
+    }
+
+    this.name = 'WarehouseError';
+    this.code = code;
   }
-
-  this.name = 'WarehouseError';
-  this.message = msg;
-  this.code = code;
 }
-
-util.inherits(WarehouseError, Error);
 
 WarehouseError.ID_EXIST = 'ID_EXIST';
 WarehouseError.ID_NOT_EXIST = 'ID_NOT_EXIST';
