@@ -1,22 +1,19 @@
 'use strict';
 
-const util = require('../util');
 const WarehouseError = require('../error');
 
-/**
- * PopulationError constructor
- *
- * @class
- * @param {String} msg
- * @extends WarehouseError
- */
-function PopulationError(msg) {
-  WarehouseError.call(this);
+class PopulationError extends WarehouseError {
 
-  this.name = 'PopulationError';
-  this.message = msg;
+  /**
+   * PopulationError constructor
+   *
+   * @param {string} msg
+   */
+  constructor(msg) {
+    super(msg);
+
+    this.name = 'PopulationError';
+  }
 }
-
-util.inherits(PopulationError, WarehouseError);
 
 module.exports = PopulationError;
