@@ -1,6 +1,5 @@
 'use strict';
 
-const util = require('../util');
 const WarehouseError = require('../error');
 
 /**
@@ -10,13 +9,12 @@ const WarehouseError = require('../error');
  * @param {String} msg
  * @extends WarehouseError
  */
-function ValidationError(msg) {
-  WarehouseError.call(this);
+class ValidationError extends WarehouseError {
+  constructor(msg) {
+    super(msg);
 
-  this.name = 'ValidationError';
-  this.message = msg;
+    this.name = 'ValidationError';
+  }
 }
-
-util.inherits(ValidationError, WarehouseError);
 
 module.exports = ValidationError;
