@@ -3,7 +3,7 @@
 const { EventEmitter } = require('events');
 const cloneDeep = require('lodash/cloneDeep');
 const Promise = require('bluebird');
-const { parseArgs, getProp, setGetter, shuffle, inherits } = require('./util');
+const { parseArgs, getProp, setGetter, shuffle } = require('./util');
 const Document = require('./document');
 const Query = require('./query');
 const Schema = require('./schema');
@@ -12,78 +12,79 @@ const WarehouseError = require('./error');
 const PopulationError = require('./error/population');
 const Mutex = require('./mutex');
 
-/**
- * Model constructor.
- *
- * @class
- * @param {string} name Model name
- * @param {Schema|object} [schema] Schema
- * @extends EventEmitter
- */
-function Model(name, schema_) {
-  EventEmitter.call(this);
+class Model extends EventEmitter {
 
-  let schema;
+  /**
+   * Model constructor.
+   *
+   * @param {string} name Model name
+   * @param {Schema|object} [schema] Schema
+   */
+  constructor(name, schema_) {
+    super();
 
-  // Define schema
-  if (schema_ instanceof Schema) {
-    schema = schema_;
-  } else if (typeof schema_ === 'object') {
-    schema = new Schema(schema_);
-  } else {
-    schema = new Schema();
-  }
+    let schema;
 
-  // Set `_id` path for schema
-  if (!schema.path('_id')) {
-    schema.path('_id', {type: Types.CUID, required: true});
-  }
+    // Define schema
+    if (schema_ instanceof Schema) {
+      schema = schema_;
+    } else if (typeof schema_ === 'object') {
+      schema = new Schema(schema_);
+    } else {
+      schema = new Schema();
+    }
 
-  this.name = name;
-  this.data = {};
-  this._mutex = new Mutex();
-  this.schema = schema;
-  this.length = 0;
+    // Set `_id` path for schema
+    if (!schema.path('_id')) {
+      schema.path('_id', {type: Types.CUID, required: true});
+    }
 
-  const _Document = this.Document = function(data) {
-    Document.call(this, data);
+    this.name = name;
+    this.data = {};
+    this._mutex = new Mutex();
+    this.schema = schema;
+    this.length = 0;
 
-    // Apply getters
-    schema._applyGetters(this);
-  };
+    class _Document extends Document {
+      constructor(data) {
+        super(data);
 
-  inherits(_Document, Document);
-  _Document.prototype._model = this;
-  _Document.prototype._schema = schema;
+        // Apply getters
+        schema._applyGetters(this);
+      }
+    }
 
-  const _Query = this.Query = function(data) {
-    Query.call(this, data);
-  };
+    this.Document = _Document;
 
-  inherits(_Query, Query);
-  _Query.prototype._model = this;
-  _Query.prototype._schema = schema;
+    _Document.prototype._model = this;
+    _Document.prototype._schema = schema;
 
-  // Apply static methods
-  const statics = schema.statics;
-  const staticKeys = Object.keys(statics);
+    class _Query extends Query {}
 
-  for (let i = 0, len = staticKeys.length; i < len; i++) {
-    const key = staticKeys[i];
-    this[key] = statics[key];
-  }
+    this.Query = _Query;
 
-  // Apply instance methods
-  const methods = schema.methods;
-  const methodKeys = Object.keys(methods);
+    _Query.prototype._model = this;
+    _Query.prototype._schema = schema;
 
-  for (let i = 0, len = methodKeys.length; i < len; i++) {
-    const key = methodKeys[i];
-    _Document.prototype[key] = methods[key];
+    // Apply static methods
+    const statics = schema.statics;
+    const staticKeys = Object.keys(statics);
+
+    for (let i = 0, len = staticKeys.length; i < len; i++) {
+      const key = staticKeys[i];
+      this[key] = statics[key];
+    }
+
+    // Apply instance methods
+    const methods = schema.methods;
+    const methodKeys = Object.keys(methods);
+
+    for (let i = 0, len = methodKeys.length; i < len; i++) {
+      const key = methodKeys[i];
+      _Document.prototype[key] = methods[key];
+    }
   }
 }
-
-inherits(Model, EventEmitter);
 
 /**
  * Creates a new document.
