@@ -1,8 +1,10 @@
 'use strict';
 
-function Mutex() {
-  this._locked = false;
-  this._queue = [];
+class Mutex {
+  constructor() {
+    this._locked = false;
+    this._queue = [];
+  }
 }
 
 Mutex.prototype.lock = function(fn) {
