@@ -3,15 +3,17 @@
 const Promise = require('bluebird');
 const { parseArgs, shuffle } = require('./util');
 
-/**
- * Query constructor.
- *
- * @class
- * @param {Array} data
- */
-function Query(data) {
-  this.data = data;
-  this.length = data.length;
+class Query {
+
+  /**
+   * Query constructor.
+   *
+   * @param {Array} data
+   */
+  constructor(data) {
+    this.data = data;
+    this.length = data.length;
+  }
 }
 
 /**
