@@ -7,10 +7,10 @@ const util = require('./util');
 const PopulationError = require('./error/population');
 const isPlainObject = require('is-plain-object');
 
-const getProp = util.getProp;
-const setProp = util.setProp;
-const delProp = util.delProp;
-const isArray = Array.isArray;
+const { getProp } = util;
+const { setProp } = util;
+const { delProp } = util;
+const { isArray } = Array;
 
 const builtinTypes = {
   String: true,
@@ -22,37 +22,39 @@ const builtinTypes = {
   Buffer: true
 };
 
-/**
- * Schema constructor.
- *
- * @class
- * @param {Object} schema
- */
-function Schema(schema) {
-  this.paths = {};
-  this.statics = {};
-  this.methods = {};
+class Schema {
 
-  this.hooks = {
-    pre: {
-      save: [],
-      remove: []
-    },
-    post: {
-      save: [],
-      remove: []
+  /**
+   * Schema constructor.
+   *
+   * @param {Object} schema
+   */
+  constructor(schema) {
+    this.paths = {};
+    this.statics = {};
+    this.methods = {};
+
+    this.hooks = {
+      pre: {
+        save: [],
+        remove: []
+      },
+      post: {
+        save: [],
+        remove: []
+      }
+    };
+
+    this.stacks = {
+      getter: [],
+      setter: [],
+      import: [],
+      export: []
+    };
+
+    if (schema) {
+      this.add(schema);
     }
-  };
-
-  this.stacks = {
-    getter: [],
-    setter: [],
-    import: [],
-    export: []
-  };
-
-  if (schema) {
-    this.add(schema);
   }
 }
 
