@@ -1,76 +1,76 @@
 'use strict';
 
-const util = require('./util');
+const { setProp } = require('./util');
 const ValidationError = require('./error/validation');
 
-const setProp = util.setProp;
+class SchemaType {
 
-/**
- * SchemaType constructor.
- *
- * This is the basic schema type. All schema types should inherit from this
- * class. For example:
- *
- * ``` js
- * var SchemaTypeCustom = function(name, options){
- *   SchemaType.call(this, name, options);
- * };
- *
- * require('util').inherits(SchemaTypeCustom, SchemaType);
- * ```
- *
- * **Query operators**
- *
- * To add a query operator, defines a method whose name is started with `q$`.
- * For example:
- *
- * ``` js
- * SchemaTypeCustom.q$foo = function(value, query, data){
- *   // ...
- * };
- * ```
- *
- * The `value` parameter is the value of specified field; the `query` parameter
- * is the value passed to the query operator; the `data` parameter is the
- * complete data.
- *
- * The return value must be a boolean indicating whether the data passed.
- *
- * **Update operators**
- *
- * To add a update operator, defines a method whose name is started with `u$`.
- * For example:
- *
- * ``` js
- * SchemaTypeCustom.u$foo = function(value, update, data){
- *   // ...
- * };
- *
- * The `value` parameter is the value of specified field; the `update` parameter
- * is the value passed to the update operator; the `data` parameter is the
- * complete data.
- *
- * The return value will replace the original data.
- *
- * @class
- * @param {String} name
- * @param {Object} [options]
- *   @param {Boolean} [options.required=false]
- *   @param {*} [options.default]
- */
-function SchemaType(name, options) {
-  this.name = name || '';
+  /**
+   * SchemaType constructor.
+   *
+   * This is the basic schema type. All schema types should inherit from this
+   * class. For example:
+   *
+   ``` js
+   class SchemaTypeCustom extends SchemaType {
+     constructor(name, options) {
+       super(name, options);
+     }
+   }
+   ```
+   *
+   * **Query operators**
+   *
+   * To add a query operator, defines a method whose name is started with `q$`.
+   * For example:
+   *
+   * ``` js
+   * SchemaTypeCustom.q$foo = function(value, query, data){
+   *   // ...
+   * };
+   * ```
+   *
+   * The `value` parameter is the value of specified field; the `query` parameter
+   * is the value passed to the query operator; the `data` parameter is the
+   * complete data.
+   *
+   * The return value must be a boolean indicating whether the data passed.
+   *
+   * **Update operators**
+   *
+   * To add a update operator, defines a method whose name is started with `u$`.
+   * For example:
+   *
+   * ``` js
+   * SchemaTypeCustom.u$foo = function(value, update, data){
+   *   // ...
+   * };
+   *
+   * The `value` parameter is the value of specified field; the `update` parameter
+   * is the value passed to the update operator; the `data` parameter is the
+   * complete data.
+   *
+   * The return value will replace the original data.
+   *
+   * @param {string} name
+   * @param {object} [options]
+   *   @param {boolean} [options.required=false]
+   *   @param {*} [options.default]
+   */
+  constructor(name = '', options) {
+    this.name = name;
 
-  this.options = Object.assign({
-    required: false
-  }, options);
+    this.options = Object.assign({
+      required: false
+    }, options);
 
-  const default_ = this.options.default;
+    const default_ = this.options.default;
 
-  if (typeof default_ === 'function') {
-    this.default = default_;
-  } else {
-    this.default = () => default_;
+    if (typeof default_ === 'function') {
+      this.default = default_;
+    } else {
+      this.default = () => default_;
+    }
   }
 }
 
