@@ -1,31 +1,31 @@
 'use strict';
 
 const SchemaType = require('../schematype');
-const util = require('../util');
 const ValidationError = require('../error/validation');
 
-const isArray = Array.isArray;
+const { isArray } = Array;
 
 /**
  * Array schema type.
- *
- * @class
- * @param {String} name
- * @param {Object} [options]
- *   @param {Boolean} [options.required=false]
- *   @param {Array|Function} [options.default=[]]
- *   @param {SchemaType} [options.child]
- * @extends {SchemaType}
  */
-function SchemaTypeArray(name, options) {
-  SchemaType.call(this, name, Object.assign({
-    default: []
-  }, options));
+class SchemaTypeArray extends SchemaType {
 
-  this.child = this.options.child || new SchemaType(name);
+  /**
+   *
+   * @param {String} name
+   * @param {Object} [options]
+   *   @param {Boolean} [options.required=false]
+   *   @param {Array|Function} [options.default=[]]
+   *   @param {SchemaType} [options.child]
+   */
+  constructor(name, options) {
+    super(name, Object.assign({
+      default: []
+    }, options));
+
+    this.child = this.options.child || new SchemaType(name);
+  }
 }
-
-util.inherits(SchemaTypeArray, SchemaType);
 
 /**
  * Casts an array and its child elements.
