@@ -1,24 +1,12 @@
 'use strict';
 
 const SchemaType = require('../schematype');
-const util = require('../util');
 const ValidationError = require('../error/validation');
 
 /**
  * Boolean schema type.
- *
- * @class
- * @param {String} name
- * @param {Object} [options]
- *   @param {Boolean} [options.required=false]
- *   @param {Boolean|Function} [options.default]
- * @extends {SchemaType}
  */
-function SchemaTypeBoolean(name, options) {
-  SchemaType.call(this, name, options);
-}
-
-util.inherits(SchemaTypeBoolean, SchemaType);
+class SchemaTypeBoolean extends SchemaType {}
 
 /**
  * Casts a boolean.
