@@ -1,24 +1,13 @@
 'use strict';
 
 const SchemaType = require('../schematype');
-const util = require('../util');
 const cuid = require('cuid');
 const ValidationError = require('../error/validation');
 
 /**
  * [CUID](https://github.com/ericelliott/cuid) schema type.
- *
- * @class
- * @param {String} name
- * @param {Object} [options]
- *   @param {Boolean} [options.required=false]
- * @extends {SchemaType}
  */
-function SchemaTypeCUID(name, options) {
-  SchemaType.call(this, name, options);
-}
-
-util.inherits(SchemaTypeCUID, SchemaType);
+class SchemaTypeCUID extends SchemaType {}
 
 /**
  * Casts data. Returns a new CUID only if value is null and the field is
