@@ -1,27 +1,27 @@
 'use strict';
 
 const SchemaType = require('../schematype');
-const util = require('../util');
 const ValidationError = require('../error/validation');
 
 /**
  * Enum schema type.
- *
- * @class
- * @param {String} name
- * @param {Object} options
- *   @param {Boolean} [options.required=false]
- *   @param {Array} options.elements
- *   @param {*} [options.default]
- * @extends {SchemaType}
  */
-function SchemaTypeEnum(name, options) {
-  SchemaType.call(this, name, Object.assign({
-    elements: []
-  }, options));
-}
+class SchemaTypeEnum extends SchemaType {
 
-util.inherits(SchemaTypeEnum, SchemaType);
+  /**
+   *
+   * @param {string} name
+   * @param {object} options
+   *   @param {Boolean} [options.required=false]
+   *   @param {Array} options.elements
+   *   @param {*} [options.default]
+   */
+  constructor(name, options) {
+    super(name, Object.assign({
+      elements: []
+    }, options));
+  }
+}
 
 /**
  * Validates data. The value must be one of elements set in the options.
