@@ -1,24 +1,12 @@
 'use strict';
 
 const SchemaTypeNumber = require('./number');
-const util = require('../util');
 const ValidationError = require('../error/validation');
 
 /**
  * Integer schema type.
- *
- * @class
- * @param {String} name
- * @param {Object} options
- *   @param {Boolean} [options.required=false]
- *   @param {Number|Function} [options.default]
- * @extends {SchemaTypeNumber}
  */
-function SchemaTypeInteger(name, options) {
-  SchemaTypeNumber.call(this, name, options);
-}
-
-util.inherits(SchemaTypeInteger, SchemaTypeNumber);
+class SchemaTypeInteger extends SchemaTypeNumber {}
 
 /**
  * Casts a integer.
