@@ -1,24 +1,12 @@
 'use strict';
 
 const SchemaType = require('../schematype');
-const util = require('../util');
 const ValidationError = require('../error/validation');
 
 /**
  * Number schema type.
- *
- * @class
- * @param {String} name
- * @param {Object} options
- *   @param {Boolean} [options.required=false]
- *   @param {Number|Function} [options.default]
- * @extends {SchemaType}
  */
-function SchemaTypeNumber(name, options) {
-  SchemaType.call(this, name, options);
-}
-
-util.inherits(SchemaTypeNumber, SchemaType);
+class SchemaTypeNumber extends SchemaType {}
 
 /**
  * Casts a number.
