@@ -1,24 +1,22 @@
 'use strict';
 
 const SchemaType = require('../schematype');
-const util = require('../util');
 
 /**
  * Object schema type.
- *
- * @class
- * @param {String} name
- * @param {Object} [options]
- *   @param {Boolean} [options.required=false]
- *   @param {Object|Function} [options.default={}]
- * @extends {SchemaType}
  */
-function SchemaTypeObject(name, options) {
-  SchemaType.call(this, name, Object.assign({
-    default: {}
-  }, options));
-}
+class SchemaTypeObject extends SchemaType {
 
-util.inherits(SchemaTypeObject, SchemaType);
+  /**
+   *
+   * @param {String} name
+   * @param {Object} [options]
+   *   @param {Boolean} [options.required=false]
+   *   @param {Object|Function} [options.default={}]
+   */
+  constructor(name, options) {
+    super(name, Object.assign({ default: {} }, options));
+  }
+}
 
 module.exports = SchemaTypeObject;
