@@ -1,24 +1,12 @@
 'use strict';
 
 const SchemaType = require('../schematype');
-const util = require('../util');
 const ValidationError = require('../error/validation');
 
 /**
  * String schema type.
- *
- * @class
- * @param {String} name
- * @param {Object} [options]
- *   @param {Boolean} [options.required=false]
- *   @param {String|Function} [options.default]
- * @extends {SchemaType}
  */
-function SchemaTypeString(name, options) {
-  SchemaType.call(this, name, options);
-}
-
-util.inherits(SchemaTypeString, SchemaType);
+class SchemaTypeString extends SchemaType {}
 
 /**
  * Casts a string.
