@@ -1,23 +1,12 @@
 'use strict';
 
 const SchemaType = require('../schematype');
-const util = require('../util');
-
-const setGetter = util.setGetter;
+const { setGetter } = require('../util');
 
 /**
  * Virtual schema type.
- *
- * @class
- * @param {String} name
- * @param {Object} [options]
- * @extends {SchemaType}
  */
-function SchemaTypeVirtual(name, options) {
-  SchemaType.call(this, name, options);
-}
-
-util.inherits(SchemaTypeVirtual, SchemaType);
+class SchemaTypeVirtual extends SchemaType {}
 
 /**
  * Add a getter.
